# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "BackupRequest": {
        "properties": {
          "passphrase": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Passphrase",
            "type": "string"
          }
        },
        "required": [
          "passphrase"
        ],
        "title": "BackupRequest",
        "type": "object"
      },
      "BackupRestoreRequest": {
        "properties": {
          "backup": {
            "additionalProperties": true,
            "title": "Backup",
            "type": "object"
          },
          "passphrase": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Passphrase",
            "type": "string"
          }
        },
        "required": [
          "passphrase",
          "backup"
        ],
        "title": "BackupRestoreRequest",
        "type": "object"
      },
...
      "CircumventionRequest": {
        "properties": {
          "country": {
            "default": "",
            "maxLength": 2,
            "title": "Country",
            "type": "string"
          },
          "custom_bridges": {
            "items": {
              "type": "string"
            },
            "maxItems": 40,
            "title": "Custom Bridges",
            "type": "array"
          },
          "mode": {
            "pattern": "^(direct|auto|manual)$",
            "title": "Mode",
            "type": "string"
          },
          "transport": {
            "default": "snowflake",
            "maxLength": 32,
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "mode"
        ],
        "title": "CircumventionRequest",
        "type": "object"
      },
      "ConfigImportRequest": {
        "properties": {
          "document": {
            "additionalProperties": true,
            "title": "Document",
            "type": "object"
          }
        },
        "required": [
          "document"
        ],
        "title": "ConfigImportRequest",
        "type": "object"
      },
      "ConfigurationDocument": {
        "properties": {
          "blocked_devices": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Blocked Devices",
            "type": "array"
          },
          "circumvention": {
            "additionalProperties": true,
            "title": "Circumvention",
            "type": "object"
          },
          "device_access": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Device Access",
            "type": "array"
          },
          "device_name": {
            "title": "Device Name",
            "type": "string"
          },
          "dns_filter": {
            "additionalProperties": true,
            "title": "Dns Filter",
            "type": "object"
          },
          "exported_at": {
            "title": "Exported At",
            "type": "integer"
          },
          "tor_policy": {
            "additionalProperties": true,
            "title": "Tor Policy",
            "type": "object"
          },
          "version": {
            "const": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "exported_at",
          "device_name",
          "tor_policy",
          "dns_filter",
          "blocked_devices",
          "circumvention"
        ],
        "title": "ConfigurationDocument",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SystemActionRequest": {
        "properties": {
          "action": {
            "pattern": "^[a-z-]{3,24}$",
            "title": "Action",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "SystemActionRequest",
        "type": "object"
      },
...
      "UpdateSettingsRequest": {
        "properties": {
          "apply": {
            "default": true,
            "title": "Apply",
            "type": "boolean"
          },
          "channel": {
            "pattern": "^(stable|edge)$",
            "title": "Channel",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "schedule": {
            "maxLength": 41,
            "minLength": 5,
            "title": "Schedule",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "schedule"
        ],
        "title": "UpdateSettingsRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OnionPi API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/circumvention": {
      "get": {
        "operationId": "circumvention_state_api_v1_circumvention_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Circumvention State Api V1 Circumvention Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Circumvention State"
      },
      "post": {
        "operationId": "update_circumvention_api_v1_circumvention_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircumventionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Circumvention Api V1 Circumvention Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Circumvention"
      }
    },
...
    "/api/v1/logs": {
      "get": {
        "operationId": "logs_api_v1_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "service",
            "required": false,
            "schema": {
              "default": "tor",
              "pattern": "^(tor|NetworkManager|dnsmasq|nftables|onionpi-firewall|onionpi|snowflake-proxy)$",
              "title": "Service",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Logs Api V1 Logs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logs"
      }
    },
...
    "/api/v1/system/action": {
      "post": {
        "operationId": "run_system_action_api_v1_system_action_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run System Action Api V1 System Action Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run System Action"
      }
    },
...
    "/api/v1/system/backup": {
      "post": {
        "operationId": "create_encrypted_backup_api_v1_system_backup_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Encrypted Backup"
      }
    },
    "/api/v1/system/backup/preview": {
      "post": {
        "operationId": "preview_encrypted_backup_api_v1_system_backup_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Encrypted Backup Api V1 System Backup Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Encrypted Backup"
      }
    },
    "/api/v1/system/backup/restore": {
      "post": {
        "operationId": "restore_encrypted_backup_api_v1_system_backup_restore_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Restore Encrypted Backup Api V1 System Backup Restore Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Encrypted Backup"
      }
    },
    "/api/v1/system/config": {
      "get": {
        "operationId": "export_configuration_api_v1_system_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationDocument"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Export Configuration"
      },
      "post": {
        "operationId": "import_configuration_api_v1_system_config_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Configuration Api V1 System Config Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Configuration"
      }
    },
...
    "/api/v1/system/update/check": {
      "post": {
        "operationId": "check_for_update_api_v1_system_update_check_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Check For Update Api V1 System Update Check Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check For Update"
      }
    },
    "/api/v1/system/update/run": {
      "post": {
        "operationId": "run_update_api_v1_system_update_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Update Api V1 System Update Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Update"
      }
    },
    "/api/v1/system/update/settings": {
      "post": {
        "operationId": "save_update_settings_api_v1_system_update_settings_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Update Settings Api V1 System Update Settings Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Update Settings"
      }
    },
...
  }
}